Task list and CYA updates
Task list links have IDs that allow the user to be focused back to where they were on the task list after answering a question.
Change links on 'Check your answers' now link to the relevant questions.
Added a 'jump' link to set up variables and take the user to the "Check your answers" page with answers.